requirements.txt
asciidoc==10.2.0
asgiref==3.5.2
autopep8==2.0.2
Babel==2.12.1
certifi==2022.12.7
charset-normalizer==3.1.0
distlib==0.3.6
dj-database-url==2.0.0
Django==4.1.4
django-money==3.1.0
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==15.3.4
filelock==3.11.0
gunicorn==20.1.0
idna==3.4
markdown2==2.4.8
pipenv==2023.3.20
platformdirs==3.2.0
py-moneyed==2.0
pycodestyle==2.10.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.6.2
tzdata==2023.2
urllib3==2.0.3
virtualenv==20.21.0
virtualenv-clone==0.5.7
whitenoise==6.4.0
